rejigger how calamares works
calamares has real trouble implementing the mount lineup we need, this
resulted in a fairly subpar installation result. we'll now sidestep
calamares for mount and unpacking use.

_calamares@subvol now takes care of both use cases. it first undoes a
bunch of stuff calamares does implicitly, then sends the rootfs subvol
to the target, creates all the necessary subvolumes and folders, mounts
/proc /dev etc and finally invokes the overlay program to overlay
everything like we do in the initrd

as a side effect the overlay program now accepts device and mount target
as input with default values when no input is given (i.e. the initrd
scenario - though now that I think about it we could probably just
invoke it properly from the service file)

the end result of all this is a more correct target partition with
properly populated /etc (overlay)