move to /var/lib/snapd/snap
as it turns out the AUR build of snapd explicitly builds with a
different path for snaps. to make things work that way and match snapd
expectations we now mount the snap volume into /var/lib/snapd/snap
(which is created during rootfs creation).
we also symlink /snap there so to the user everything seems as though it
was under /snap

if that doesn't work out we might have to reconsider using the AUR
package :\