live: change the way flatpak setup works
instead of seeding the flatpaks manually let's just use the tree from
installation. we'll put that into /@flatpak and then mount it into place
as part of the live generator.

we'll see how it goes.

if this works out we may then btrfs receive the volume into the
installed system and handle it similar to @snap as a special subvol