Revert "Use pre-built AUR binaries instead of rebuilding them on each run"
This reverts commit 5d37d0a89ff3e45c20e3f129c7e325f83ed4b2a9.

So, while trying to set this up on KDE infra I hit the rather awkward
problem that aurto doesn't really want to be automated. It's very much
meant to be operated by a human. That is really just a source for failure.
If we want to go down this route we'll want to use aurutils directly and
write some higher level tooling on top if necessary.
Undoing this for now so we can ponder it some more.

Realistically I'd rather explore factoring out the rootfs build
into its own job though. That way we can isolate all the binary builds
into a job that produces a tarball which can then be consumed to render
an image without having to build anything. Seems to me that should rock
the boat a bit less while moving in the general direction of splitting
things out.