shrink btrfs to the actually necessary size
this doesn't yet give any gains because partition resizing also needs
figuring out, but it's a start.

specifically we now deduplicate files, balance extents and shrink the
actual filesystem as much as possible.

in a second step we could then chop off the end of the partition table
to actually reduce the image size. except it is more difficult than one
might think. perhaps we should construct a new table entirely instead.
something to figure out