Prevent Calamares from being installed/visible after installing system
We install all packages in the `kde-linux` repo:
```bash
  # All the KDE we plan to include in the base image
  $(pacman --sync --list --quiet kde-linux)
```

I recently added `calamares-git` (and some other AUR packages) to that
repo so we don't have to build them as part of every image build. This
is causing it to get included in the base image.

This change makes pacman restrict to just the `kde-linux` group instead
of the whole repo.

Resolves #42