Make zsh the default shell for new users
ZSH is (in many ways) the successor to BASH.

It looks/feels the same as bash, but with better tab-completion, better history management, typo-correction, etc.
People that care about their shell will customize it, but people that don't will just notice that the terminal in KDE Linux is faster/more responsive. We should aim to give the best user experience by default.

https://medium.com/@mattwinmill88/what-even-is-zsh-and-why-is-everyone-using-it-9695c869fbe4#0747

Note: This change only affects new accounts so it won't suddenly break anyone's setup.