erofs: prototype
requires kde-linux.erofs=1 to opt in.
expects a kde-linux.erofs in the root of the root partition.

the biggest change here is a new mount generator that replaces the old
overlay tech. to assit in migration a transition script is invoked as
part of the mount actions. all of this happens in the initrd.
after switch into the real root the most important part is the tmpfiles
system which should make sure to populate etc on clean installations

assuming the prototype works out this would render /usr readonly while
the rest of the system can just be a plain old btrfs