Remove pacman
We don't want people finding and using it anyway, and it isn't
even functional because we don't include a /etc/pacman.conf file. It
was only ever intended to be here for local debugging during bring-up,
but the fact thst it hasn't worked shows that this isn't very important
either. Let's delete it.

Fixes #161