Revert "Use UTC date with hyphens as the version string"
This reverts commit a506b2e8a8ec77161b4b38a85edddff39e706a74.

This turned out a bit problematic because systemd has no good facilities
to allow changing the version scheme. After some pondering I think
it's best if we stick to the existing number format and consider
it an internal build number that shouldn't mean anything to the outside.
The pretty versions we present to the outside can be different.