simplify the btrfs cleanup
the previous approach wasn't working correctly because paths are from
the host and mean nothing inside the container. as it turns out the
builtin recursive delete works as intended though, so let's use that.

since we need to find our "root" subvolume path now we do a simple glob
run to locate it