Re-add toolbox to base image
Add Toolbox to the base image. Its the base of distrobox
but is still useful, as its faster and lighter etc. It's
also needed for https://invent.kde.org/system/kontainer
Second Backend. It was replaced by distrobox earlier but
toolbx has been rewritten in an actual programming
language and is useful again.