Make PAM less authoritarian
The defaults are currently very harsh: each failed authentication
attempt makes you wait for 3 seconds, and 3 failed attempts lock you
out of the system for 10 minutes.

These defaults are tuned for high security, but Bugzilla is full of
complaints about it from Arch users who placed the blame on KDE. These
defaults are not well-tuned for normal users in normal environments, so
let's soften them a bit to offer a less harsh experience.

Resolves #222