btrfs-migrator: run on actually composed overlays
bit awkward but this should produce more reliably results. in particular
for flatpak the previous naive approach could easily fail because of the
ostree repo content being files or dirs depending on unknown factors,
making it super difficult to copy them reliably. by composing first we
defer that problem to overlayfs, which is how this was used in rootfsv1
anyway.

with this we also no longer need the erofs /usr mount