update protection logic
the previous code would end up dropping the protected versions from the
shasums, we'll want to keep them there as well.

to achieve this we now collect all protected releases and then append
them to the keep list once the keep list has been pruned. seems the most
reliable way of doing this