Add and pre-enable a systemd service for powertop --auto-tune
On my system this saves about 0.4W while on battery power.

The major problem with `powertop --auto-tune` in general is that it
turns on USB auto-suspend, which can wreak havoc with input devices. To
prevent that, we then undo it for detected USB input devices.

This is a workaround for an upstream issue in Powertop, but it's a very
old issue and doesn't look like it's getting fixed anytime soon. See:

- https://github.com/fenrus75/powertop/issues/37
- https://github.com/fenrus75/powertop/issues/38
- https://github.com/fenrus75/powertop/issues/101
- https://github.com/fenrus75/powertop/issues/165