mtimer: links need special handling
I completely ignored the fact that symlinks can point to directories.
this then would make our checksum logic fall over because you can't
checksum a dirent. instead introduce completely bespoke handling for
symlinks. it's increeeedibly similar to files but ever so slightly
different so we have type assurances.

may be worth creating an Analyzer interface in the future so we don't
have to dupe things so much