Use BFQ I/O scheduler for spinning disks
It offers much better latency in exchange for only a modest reduction
in throughput, and only for some workflows. High latency being a
defining negative characteristic of spinning disks, this seems like a
worthwhile trade-off for us to make for desktop usage.

This doesn't change the default I/O schedulers for flash-based storage,
as the defaults seem sensibly chosen (`none` for NVMe SSDs and
`mq-deadline` for everything else).

Related to #258