Disable audio power saving when plugged in to stop audio buzzing
The internet is absolutely chock-full of complaints about this, all
being pointed to the same workaround: disable power management for the
snd_hda_intel module. It appears to be quite widespread over a long
period of time.

Let's do it automatically so users of affected hardware don't get driven
mad.

Related to #258