Prevent tuned from messing with the vm.swappiness parameter we set
Some of the profiles — most notably "throughput-performance" which gets
activated as the "Performance mode" exposed in Plasma — will mess with
the vm.swappiness value. But we already set it to a value of our
choosing, and that value should be fine for all modes given our zram
setup. So let's stop tuned from doing this.