Tweak kernel parameters to optimize latency and power usage
See inline comments regarding what exactly these kernel options do.

But in a nutshell, we should get lower latency and power usage, at the
cost of a little bit of throughput in CPU-intensive tasks. For the kinds
of userspace desktop use cases we target in KDE Linux, this seems like
an appropriate trade-off.

Research sources included:
- https://discourse.ubuntu.com/t/enable-low-latency-features-in-the-generic-ubuntu-kernel-for-24-04/42255
- https://discourse.ubuntu.com/t/fine-tuning-the-ubuntu-24-04-kernel-for-low-latency-throughput-and-power-efficiency/44834
- https://www.kernel.org/doc/html/v6.15-rc3/timers/no_hz.html