Add support for Yubikeys
This adds support for using a Yubikey for unlocking a FDE partition or
being a second factor of authentication for anything using PAM.

Also includes is the `ykman` command-line tool for configuring and
managing Yubikeys. The "Yubikey Authenticator" GUI app is not
pre-installed, but is available on Flathub for download as needed.

All of this adds about 5.5 MB to the image.