Pre-install more filesystem support for Partition Manager
Otherwise it won't be able to handle exFat or XFS volumes, or resize FAT
volumes.

e2fsprogs is already pre-installed because it's a dependency of
something else, but let's mention it here because we don't want it to be
removed by accident if dependencies change in the future.

All of this adds 5.5 MB to the image, almost all of it from XFS support.